Move pipline logic to separate helper file in order to add more pipelines later
